We can now verify that the current Logged in is the one trying to delete a todo

+++ Module2/todoapp-api/routes/todosRouter.js
@@ -1,11 +1,12 @@
 const express = require('express')
 const TodoModel = require('../models/todoSchema')
+const authMiddleware = require('../middleware/authMiddleware')
 
 // * Create a Router
 const router = express.Router()
 
 //* GET TODOS
-router.get('/', async (req, res) => {
+router.get('/', authMiddleware, async (req, res) => {
    try {
        const todos = await TodoModel.find()
        res.status(200).json(todos)
@@ -15,8 +16,9 @@ router.get('/', async (req, res) => {
 })
 
 //* CREATE TODOS
-router.post('/', async (req, res) => {
+router.post('/', authMiddleware, async (req, res) => {
     const todoData = req.body // gets the data from the request
+    todoData.user = req.user.id
     console.log(todoData);
     try {
         const todo = await TodoModel.create(todoData) // create the todo in the db
@@ -31,9 +33,9 @@ router.post('/', async (req, res) => {
 
 
 //* GET TODO BY ID
-router.get('/:id', async (req, res) => {
+router.get('/:id', authMiddleware, async (req, res) => {
     const id = req.params.id
-
+    console.log(req.user);
     try {
         const todo = await TodoModel.findById(id)
         res.status(200).json(todo)
@@ -47,7 +49,7 @@ router.get('/:id', async (req, res) => {
 
 
 //* UPDATE TODO BY ID
-router.put('/:id', async (req, res) => {
+router.put('/:id',authMiddleware, async (req, res) => {
     const id = req.params.id
     const newTodoData = req.body
      try {
@@ -60,12 +62,20 @@ router.put('/:id', async (req, res) => {
 })
 
 //! DELETE A TODO
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', authMiddleware, async (req, res) => {
     const id = req.params.id
-
+    console.log('FROM DELETE', req.user);
     try {
+        const todoToDelete = await TodoModel.findById(id)
+        console.log(todoToDelete);
+        console.log(todoToDelete.user._id.toString(), '||', req.user.id);
+        
+        if (todoToDelete.user._id.toString() !== req.user.id){
+            return res.status(400).json({msg: 'Not Authorized!'})
+        }
         const todo = await TodoModel.findByIdAndDelete(id)
-        res.status(200).json({msg: 'Todo was deleted!'})
+        res.status(200).json('Todo deleted by the user!')
+
     } catch (error) {
         console.log(error);
     }
